fix various bottlenecks to > 65536 temps

+++ src/occopt/ilazy.h
@@ -25,8 +25,8 @@
 
 namespace Optimizer
 {
-extern unsigned short* termMap;
-extern unsigned short* termMapUp;
+extern unsigned * termMap;
+extern unsigned * termMapUp;
 extern unsigned termCount;
 
 void SetunMoveableTerms(void);
